requirements/base.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.3.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-taggit==5.0.1
django-autoslug==1.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
whitenoise==6.7.0
gunicorn==22.0.0
django-storages==1.14.3
boto3==1.34.144
requests==2.32.3
python-dateutil==2.9.0
